+++ Client/Electron/src/actions/auth0/config.js
@@ -1,0 +1,6 @@
+export const AUTH0_DOMAIN = 'react-universal.auth0.com';
+export const AUTH0_CLIENT_ID = '3821inroTsXub5iKOew7JHTT-OaJl3jp';
+export const AUTH0_CLIENT_SECRET = '3usxeq-Ae3jur3LYVlhJ0Ldjyi01sCoHsbTdRpbFLFwzeRMZjQHCp8La1ixDx9jW';
+export const AUTH0_API_AUDIENCE = 'https://api.react-universal.com';
+export const AUTH0_LOGIN_REDIRECT_URL = 'http://localhost:3000/callback';
+export const AUTH0_LOGOUT_REDIRECT_URL = 'http://localhost:3000/';
