+++ Client/Electron/src/actions/userActions.js
@@ -1,5 +1,3 @@
-import { AUTH0_DOMAIN } from '../config';
-
 export const fetchUserPending = () => {
     return {
         type: 'FETCH_USER_PENDING'
@@ -17,28 +15,4 @@ export const fetchUserError = () => {
     return {
         type: 'FETCH_USER_ERROR'
     }
-}
-
-export const fetchUser = (accessToken) => (dispatch) => {
-
-    dispatch(fetchUserPending());
-
-    fetch(`https://${AUTH0_DOMAIN}/userinfo`, {
-        headers: {
-            Authorization: `Bearer ${accessToken}`
-        }
-    })
-    .then(res => {
-        if (res.status !== 200)
-            throw `failed to get user profile with status ${res.status}`;
-        return res.json();
-    })
-    .then(user => {
-        dispatch(fetchUserSuccess(user));
-    })
-    .catch(error => {
-        console.log(error);
-        dispatch(fetchUserError());
-    })
-
 }
