+++ Client/Electron/src/components/Profile/index.js
@@ -1,8 +1,8 @@
 import Profile from "./component";
 import { connect } from "react-redux";
 import { bindActionCreators } from 'redux';
-import { fetchUser } from '../../actions/userActions';
-import { logout, removeSession } from '../../actions/authActions';
+import { logout, fetchUser } from '../../actions/auth0';
+import { removeSession } from '../../actions/authActions';
 
 const mapStateToProps = (state) => {
